separate specs into different suites
Motivation:
Instead of having all test in a single suite like we currently have we'd
like to separate them into logically grouped functionality.
We aslo would like to be able to run each spec individually to allow for
quicker development/debugging.

Modifications:
Removed index.html so that it is now genereated from a template.

Result:
Single specs be run by using:
$ npm t -- --filter=syncCrudSpec.js

JIRA:
https://issues.jboss.org/browse/FH-3155
