+++ .gitignore
@@ -1,2 +1,5 @@
 node_modules
 main.js
+.grunt
+index.html
+index.html?url=http:/
