+++ tasks/Runner.tmpl
@@ -1,0 +1,22 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>Jasmine Spec Runner</title>
+  <link rel="shortcut icon" type="image/png" href="<%= favicon %>">
+<% css.forEach(function(style){ %>
+  <link rel="stylesheet" type="text/css" href="<%= style %>">
+<% }) %>
+
+</head>
+<body>
+<% with (scripts) { %>
+  <% [].concat(polyfills, jasmine, boot, vendor, helpers, src, specs,reporters).forEach(function(script){ %>
+  <script src="<%= script %>"></script>
+  <% }) %>
+<% }; %>
+  <script>
+    jasmine.DEFAULT_TIMEOUT_INTERVAL = 15000;
+  </script>
+</body>
+</html>
